Move shared controller and secret utils to secret.go
... because they will be shared between the Service and
StatefulSet controllers.

Only moves code with no changes, should not change behavior.
